Add missing peer-dependency (`ngx-cookie`) of `@harbor/ui` (#389)
* Remove `angular-font-awesome` from lockfile

The `angular-font-awesome` project was removed from dependencies
in #384, but `yarn.lock` was not updated to reflect that.

* Avoid updating ChromeDriver twice

ChromeDriver is updated in the `postinstall` npm script. There is no
need to update it again in CI, as long as we specify the correct
ChromeDriver version to download in the `postinstall` script.

* Ensure `passing.json` only contains existing projects

Previously, when a project was removed from `angular.json`, it might
still remain in `passing.json`. As a result, it was difficult to
identify how many projects are actually known to be passing by looking
at `passing.json`.

This commit adds an additional check that fails if `passing.json`
contains a project that is not in `angular.json` (and thus not tested
against any more).

* Switch from listing passing projects to listing failing ones

Previously, there was list of passing projects that was used to guard
against regressions. Now that more projects are passing rather than
failing, it makes more sense to track failing projects instead.

* Add back `angular-tree-component` and `mydatepicker`

These two projects were removed in d845858b8 and e7408c70b to speed up
CI. They might be passing at the time, but they are certainly failing
now.

This commit adds them back, so we can track and fix the failures (and
prevent future regressions).

* Add missing peer-depenency (`ngx-cookie`) of `@harbor/ui`

Previously, `ngx-cookie` used to be install as a direct dependency for
the `ngx-cookie-ngcc` project. In d845858b8, the project was removed and
so was the `ngx-cookie` dependency. However, it turns out that
`ngx-cookie` is also a peer-dependency of `@harbor/ui`, so we still need
it as a dependency.

This commit adds it back to allow `@harbor/ui` compilation to advance
until the next (unrelated) error:

```
node_modules/@harbor/ui/fesm5/harbor-ui.js(15723,30):
  error TS-991010: Value at position 7 in the NgModule.imports of SharedModule is not a reference: [object Object]
```