build(deps): change references to GitHub commits to use shorthand
Before #1404, the `github:<org>/<repo>#<commit>` format was used for
installing Angular packages from `*-builds` repos. In #1411, we switched
back to GitHub commits as versions, but with a different format:
`https://github.com/<org>/<repo>#<commit>`.

This commit switches back to the short format, because:
a. It is shorter.
b. It might be related to Renovate's failing to update Angular framework
   and CLI recently (e.g. #1468 and #1469).