refactor: change `@angular/platform-server` project/test to be more realistic
Previously, the project used to verify the `@angular/platform-server`
package's compatibility with ngcc was importing
`@angular/platform-server` into the app (which is not what would happen
in an actual app).

This commit changes the project to look more like a real-world app. It
also changes the tests to make sure they better exercises the SSR
functionality.