build: mark projects as failing to fix tests with latest Angular framework
In the latest commits of the Angular frameowork (targeting v11.0.0),
there were some minor breaking changes that broke two packages used as
dependencies in the `ngcc-validation` repo:
- `@delon/theme`
- `igniteui-angular`

As a result, the sub-projects depending on these packages started
failing on CI with the latest builds of Angular framework.

This commit adds the affected projects to the list of failing projects
to allow updating to the latest Angular framework and be able to catch
any future regressions asap.

NOTE:
PRs with fixes have been submitted to both packages (ng-alain/delon#1054
and IgniteUI/igniteui-angular#8239) and the corresponding projects will
be removed from the list of failing projects as soon as the dependencies
have been upgrade to versions including the fixes.