test: fail tests if the failing projects list contains passing projects
Previously, if the tests for a project that was marked as failing (i.e.
was included in `failing-projects.json`) started passing, this would be
logged on CI but no further action had to be taken.

This had the following undesirable effects:
- At some point, someone realized that a previously failing project was
  reported as passing on CI and updated the failing projects lists.
  However, it was very cumbersome to find out exactly what caused the
  project to be fixed. (Was it a fix in the Angular framework? Was the
  error fixed in the upstream package? Was it fixed in one of its
  dependencies?)
- If a failing project was fixed but then later started failing for a
  different reason (while still being included in the list of failing
  projects), the tests on CI would still pass and we would not get
  notified about the new failure.

This commit changes this behavior so that tests will fail if a project
marked as failing starts passing. To make the tests pass, one has to remove the
affected project from the list of failing projects.