build: update dependencies and switch to pinned dependencies
It seems that by default Renovate will not update a dependency if the
latest version falls within the version range specified in
`package.json`. We had several dependencies with `^` ranges, which meant
that Renovate would only update them on the next major version. As a
result, several dependencies were not on the latest version.

This commit updates all dependencies to latest versions. It also ensures
all dependencies in `package.json` use a specific version (and not a
version range), which will allow Renovate to update them even on patch
releases. Finally, it sets [rangeStrategy][1] to `pin` to make sure we
keep pinned dependencies and continue to update all dependencies for any
type of release in the future.

Internal discussion:
https://angular-team.slack.com/archives/CB4UC1932/p1614610072009900

NOTE
A couple of dependencies were intentionally kept back:
- `pdfjs-dist` (a peer dependency of `@alfresco/adf-core`) was kept back
   at version 2.5.207. `@alfresco/adf-core` does not seem to be
   compatible with more recent versions of `pdfjs-dist` (v2.6.x).
- `@types/googlemaps` (a peer dependency of `@agm/core`) was kept back
  at version 3.39.13. `@agm/core` is not compatible with more recent
  versions of `@types/googlemaps` (v3.39.14+).

[1]: https://docs.renovatebot.com/configuration-options/#rangestrategy