build: remove the `angular-router-yg-ngcc` project
As seen on its [NPM page][1], the `angular-router-yg` package has less
than 200 weekly downloads, has not been updated in 4 years and its
source seems to be the `@angular/router` package.

This package is not compatible with RxJS v7, because it imports from
some removed entry-points (see [example failure][2]) and is unlikely to
be updated at this point.

Since there seems to be little value in testing against this package,
this commit removes the corresponding `angular-router-yg-ngcc` project
(opening the way to updating RxJS to v7).

[1]: https://www.npmjs.com/package/angular-router-yg
[2]: https://circleci.com/gh/angular/ngcc-validation/53330