build: group Angular framework and CLI updates together to avoid version incompatibility
When the Angular framework releases a new major version, that version is
only compatible with the same major version of Angular CLI and vice
versa.

With the previous config, Renovate would create two separate PRs, one to
update the Angular framework to the latest version and one for Angular
CLI. Both PRs would be failing tests however and they needed to be
manually combined to pass CI (so that both major version align).

This commit removes the need for manual intervention when a new major
version is released by configuring Renovate to combine updates to the
Angular framework and CLI. The downside is that if such a PR is broken
_and_ it contains updates to both framework and CLI, it might not be
obvious from the CI logs, which is the culprit. This is a very minor
inconvenience though and the probability of it happening is very low.